fix: fix math with timeline and commit counts of a PR (#19)
* fix: fix math with timeline and commit counts of a PR

From our Enterprise Server at 3.4.2, we found cases where a pull request has
more than 250 commits. It seems like the timeline count doesn't incorporate the
full commit count, so the subtraction of commit count from the timeline count
would result in a negative total record count.

This commit changes the math of the issue event count so that only 250 commits
are subtracted from the timeline count if the commit count is greater than 250.
This logic and behavior of the timeline GraphQL property needs to be confirmed
by a GitHub engineer.

NOTE: Similar changes may need to be made with respect to the comment count of
a pull request as well. I just haven't found a use case to test it myself.

* style: fix indentation of warning message